refactoring: 3차 리펙터링  - model -> entity 패키지 변경
