+++ src/main/java/com/api/onboardingkit/article/repository/HashtagRepository.java
@@ -1,6 +1,6 @@
 package com.api.onboardingkit.article.repository;
 
-import com.api.onboardingkit.article.model.Hashtag;
+import com.api.onboardingkit.article.entity.Hashtag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
