+++ src/main/java/com/api/onboardingkit/article/service/ArticleService.java
@@ -4,8 +4,8 @@ import com.api.onboardingkit.article.repository.ArticleRepository;
 import com.api.onboardingkit.article.repository.HashtagRepository;
 import com.api.onboardingkit.article.dto.ArticleRequestDTO;
 import com.api.onboardingkit.article.dto.ArticleResponseDTO;
-import com.api.onboardingkit.article.model.Article;
-import com.api.onboardingkit.article.model.Hashtag;
+import com.api.onboardingkit.article.entity.Article;
+import com.api.onboardingkit.article.entity.Hashtag;
 import com.api.onboardingkit.article.dto.ArticleSearchDTO;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
